css/css-scroll-snap/parsing/scroll-padding-left-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Scroll Snap Test: scroll-padding-left with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-scroll-snap-1/#propdef-scroll-padding-left">
<meta name="assert" content="scroll-padding-left supports the full grammar 'auto | <length-percentage>">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("scroll-padding-left", "auto");

test_valid_value("scroll-padding-left", "10px");
test_valid_value("scroll-padding-left", "20%");
test_valid_value("scroll-padding-left", "calc(2em + 3ex)");
</script>
</body>
</html>
